# Conflicts:
#	applications/GeoMechanicsApplication/custom_constitutive/incremental_linear_elastic_law.cpp
#	applications/GeoMechanicsApplication/custom_processes/set_absorbing_boundary_parameters_process.hpp
#	applications/GeoMechanicsApplication/custom_python/add_custom_strategies_to_python.cpp
#	applications/GeoMechanicsApplication/custom_strategies/schemes/README.md

+++ applications/ChimeraApplication/custom_utilities/hole_cutting_utility.cpp
@@ -55,7 +55,7 @@ void ChimeraHoleCuttingUtility::RemoveOutOfDomainElements(
         for (auto& node : geom) {
             double nodal_distance = node.FastGetSolutionStepValue(CHIMERA_DISTANCE);
 
-            nodal_distance = nodal_distance * DomainType;
+            nodal_distance = nodal_distance * static_cast<double>(DomainType);
             if (nodal_distance < -1 * OverLapDistance) {
                 ++num_nodes_outside;
                 is_elem_outside = is_elem_outside && true;
